renamed Server.java because it was colliding with commands.Server... the cli server command now starts a ListenServer
