+++ src/main/java/ch/heigvd/dai/commands/Server.java
@@ -4,6 +4,8 @@ import ch.heigvd.dai.server.ListenServer;
 import java.util.concurrent.Callable;
 import picocli.CommandLine;
 
+import ch.heigvd.dai.server.ListenServer;
+
 @CommandLine.Command(name = "server", description = "Start the server part of the network game.")
 public class Server implements Callable<Integer> {
 
@@ -15,7 +17,7 @@ public class Server implements Callable<Integer> {
 
   @Override
   public Integer call() {
-    ListenServer.run(port, 10);
+    ListenServer.run(port,10);
     return 0;
   }
 }
