+++ src/main/java/ch/heigvd/dai/server/ListenServer.java
@@ -1,23 +1,24 @@
 package ch.heigvd.dai.server;
 
-import ch.heigvd.dai.gameclass.ServerPlayer;
+import ch.heigvd.dai.gameclass.Player;
+
 import java.io.IOException;
 import java.net.ServerSocket;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
+import java.util.LinkedList;
 
 public class ListenServer {
-  public static void run(final int PORT, final int MAX_GAMES) {
-    System.out.println("[SERVEUR] : Starting");
+    private static LinkedList<Player> players = new LinkedList<>();
+    public static void run(final int PORT, final int MAX_GAMES){
+        System.out.println("[SERVEUR] : Starting");
 
-    try (ServerSocket serverSocket = new ServerSocket(PORT);
-        ExecutorService executor = Executors.newFixedThreadPool(MAX_GAMES * 3)) {
-      final GameManager manager = new GameManager();
-      while (true) {
-        executor.submit(new ServerPlayer(serverSocket.accept(), manager));
-      }
-    } catch (IOException e) {
-      System.err.println("[SERVEUR] : " + e.getMessage());
+        final GameManager manager = new GameManager(MAX_GAMES);
+
+        try(ServerSocket serverSocket = new ServerSocket(PORT)){
+            players.add(new Player(serverSocket.accept(),manager));
+        }catch (IOException e){
+            System.err.println("[SERVEUR] : " + e.getMessage());
+        }
+
+        players.removeIf(Player::isDone);
     }
-  }
 }
